+++ website/examples/comparisons_reverse.py
@@ -4,7 +4,7 @@ import dysonsphere as ds
 
 ds.theme()
 
-cars = ds.ensure_polars(data.cars()).drop_nulls(["Miles_per_Gallon"])
+cars = ds.utils.ensure_polars(data.cars()).drop_nulls(["Miles_per_Gallon"])
 origins = ["Europe", "Japan", "USA"]
 pairs = [("Europe", "Japan"), ("Europe", "USA")]